Users in DB

+++ project.py
@@ -116,6 +116,11 @@ def gconnect():
     login_session['picture'] = data['picture']
     login_session['email'] = data['email']
 
+    userId = getUserID(login_session['email'])
+    if userId == None:
+        userId = createUser(login_session)
+    login_session['user_id'] = userId
+
     output = ''
     output += '<h1>Welcome, '
     output += login_session['username']
@@ -187,7 +192,11 @@ def restaurantsJSON():
 @app.route('/restaurant/')
 def showRestaurants():
     restaurants = session.query(Restaurant).order_by(asc(Restaurant.name))
-    return render_template('restaurants.html', restaurants=restaurants)
+    print login_session['email']
+    print getUserID(login_session['email'])
+    if 'username' in login_session:
+      return render_template('restaurants.html', restaurants=restaurants)
+    return render_template('publicrestaurants.html', restaurants=restaurants)
 
 # Create a new restaurant
 @app.route('/restaurant/new/', methods=['GET', 'POST'])
@@ -195,7 +204,8 @@ def newRestaurant():
     if 'username' not in login_session: 
         return redirect(url_for('showLogin'))
     if request.method == 'POST':
-        newRestaurant = Restaurant(name=request.form['name'])
+        newRestaurant = Restaurant(name=request.form['name'], 
+            user_id=login_session['user_id'])
         session.add(newRestaurant)
         flash('New Restaurant %s Successfully Created' % newRestaurant.name)
         session.commit()
@@ -226,6 +236,8 @@ def deleteRestaurant(restaurant_id):
         return redirect(url_for('showLogin'))
     restaurantToDelete = session.query(
         Restaurant).filter_by(id=restaurant_id).one()
+    if restaurantToDelete.user_id != login_session['user_id']:
+        return '<script>function myFunction() {alert("You are not authorized to delete this restaurant. Please create your own restaurant in order to delete.");}</script><body onload="myFunction()"">'    
     if request.method == 'POST':
         session.delete(restaurantToDelete)
         flash('%s Successfully Deleted' % restaurantToDelete.name)
@@ -241,8 +253,10 @@ def showMenu(restaurant_id):
     restaurant = session.query(Restaurant).filter_by(id=restaurant_id).one()
     items = session.query(MenuItem).filter_by(
         restaurant_id=restaurant_id).all()
-    return render_template('menu.html', items=items, restaurant=restaurant)
-
+    creator = getUserInfo(restaurant.user_id)
+    if login_session['user_id'] == restaurant.user_id:
+        return render_template('menu.html', items=items, restaurant=restaurant, creator = creator)
+    return render_template('publicmenu.html', items=items, restaurant=restaurant, creator = creator)
 
 # Create a new menu item
 @app.route('/restaurant/<int:restaurant_id>/menu/new/', methods=['GET', 'POST'])
@@ -251,8 +265,9 @@ def newMenuItem(restaurant_id):
         return redirect(url_for('showLogin'))
     restaurant = session.query(Restaurant).filter_by(id=restaurant_id).one()
     if request.method == 'POST':
-        newItem = MenuItem(name=request.form['name'], description=request.form['description'], price=request.form[
-                           'price'], course=request.form['course'], restaurant_id=restaurant_id)
+        newItem = MenuItem(name=request.form['name'], description=request.form['description'], 
+            price=request.form['price'], course=request.form['course'], restaurant_id=restaurant_id, 
+            user_id=login_session['user_id'])
         session.add(newItem)
         session.commit()
         flash('New Menu %s Item Successfully Created' % (newItem.name))
@@ -300,9 +315,22 @@ def deleteMenuItem(restaurant_id, menu_id):
         return render_template('deleteMenuItem.html', item=itemToDelete)
 
 
+def getUserID(email):
+  try:
+    user = session.query(User).filter_by(email=email).one()
+    return user.id
+  except:
+    return None
+
+
+def getUserInfo(user_id):
+    user = session.query(User).filter_by(id=user_id).one()
+    return user
+
+
 def createUser(login_session):
     newUser = User(name = login_session['username'], email =
-        login_session['email'], picture=login_session[picture])
+        login_session['email'], picture=login_session['picture'])
     session.add(newUser)
     session.commit()
     user = session.query(User).filter_by(email=login_session['email']).one()
